Added hint about using as.numeric for arithmetic challenge in novice R lesson 06.

+++ novice/r/arith.R
@@ -10,9 +10,9 @@ main <- function() {
   #   Rscript arith.R 3 - 4
   #
   args <- commandArgs(trailingOnly = TRUE)
-  num1 <- as.integer(args[1])
+  num1 <- as.numeric(args[1])
   operation <- args[2]
-  num2 <- as.integer(args[3])
+  num2 <- as.numeric(args[3])
   if (operation == "+") {
     answer <- num1 + num2
     cat(answer)
